saga_lore_module/requirements.txt
fastapi
uvicorn
chromadb
sentence-transformers
python-frontmatter
pydantic
